requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
python=3.7.8
pkgconfig=1.5.4
pytorch=1.9.1
torchaudio=0.9.1
torchvision=0.10.1
cudatoolkit=11.1.1
pillow=8.4.0
scipy=1.7.3
numpy=1.21.5
matplotlib=3.5.1
foolbox=3.3.1
tqdm=4.62.3
scikit-learn=1.0.2
cvxopt=1.2.7
cvxpy=1.1.17
cvxpy-base=1.1.17
glpk=4.65
blas=2.111
libblas=3.9.0
libcblas=3.9.0
gurobi=9.5.0